fix(eval): a line offset can be zero or negative
Ruby lets `eval` number a snippet from a line at or below one:
`eval(src, b, "f.rb", 0)` makes the first line line ZERO, which ruby
renders as a frame with no line number at all. ERB relies on it -- the
source it compiles opens with its own `#coding:UTF-8` preamble, so it
passes `0` and lets the preamble take line 0 and the template's first
line take line 1.

Zeo carried the offset as `u32`, so `0 - 1` saturated back to `0` and
every ERB frame reported one line too far down: a raise on template line
3 said `(erb):4`.

The offset is `i32` now, the whole way -- `CompileOptions`, the arena's
`SourceFile`, the lowering frame and the syntax reporter. `line_at`
sums in `i64` and floors at zero, which is not a clamp hiding an error:
zero is ALREADY the no-line marker at both layers, and it is exactly
what ruby renders for a line at or below zero.

`__LINE__` needed no change -- it already summed in `i64`, and now
answers 0 and -3 where ruby does.

+++ crates/zeo/src/parse/loader.rs
@@ -418,7 +418,7 @@ pub(super) fn lower_main_file(
                 source,
                 &result,
                 &main_name,
-                line_offset as i32 + 1,
+                line_offset + 1,
                 &ignore,
             ),
         ));
